test: migrate `get_level` tests
