+++ tests/v1_test.py
@@ -5,7 +5,7 @@ import re
 from collections import deque
 from contextlib import nullcontext as does_not_raise
 from datetime import datetime, timedelta
-from typing import TYPE_CHECKING, Any, TypeAlias, cast
+from typing import TYPE_CHECKING, Any, cast
 
 import pytest
 
@@ -420,69 +420,6 @@ def test_is_native_series(is_native_series: Callable[[Any], Any]) -> None:
     data = {"a": [1, 2]}
     ser = nw.from_native(pd.DataFrame(data))["a"]
     assert not is_native_series(ser)
-
-
-def test_get_level() -> None:
-    pytest.importorskip("polars")
-    import polars as pl
-
-    df = pl.DataFrame({"a": [1, 2, 3]})
-    assert nw_v1.get_level(nw_v1.from_native(df)) == "full"
-    assert (
-        nw_v1.get_level(
-            nw_v1.from_native(df.__dataframe__(), eager_or_interchange_only=True)
-        )
-        == "interchange"
-    )
-
-
-def test_v1_explicit_level_kwarg() -> None:
-    pytest.importorskip("polars")
-    import polars as pl
-
-    nw_lf = nw_v1.from_native(pl.LazyFrame({"a": [1]}))
-    rewrapped_lf = nw_v1.LazyFrame[pl.LazyFrame](nw_lf._compliant_frame)
-    assert nw_v1.get_level(rewrapped_lf) == "lazy"
-
-    nw_s = nw_v1.from_native(pl.Series(name="a", values=[1]), series_only=True)
-    rewrapped_s = nw_v1.Series(nw_s._compliant_series)
-    assert nw_v1.get_level(rewrapped_s) == "full"
-
-
-MainInstances: TypeAlias = tuple[nw.DataFrame[Any], nw.LazyFrame[Any], nw.Series[Any]]
-
-
-@pytest.fixture
-def main_instances(eager_implementation: EagerAllowed) -> MainInstances:
-    df = nw.DataFrame.from_dict({"a": [1, 2, 3]}, backend=eager_implementation)
-    return df, df.lazy(), df.get_column("a")
-
-
-def test_get_level_raises_main(main_instances: MainInstances) -> None:
-    df, lf, ser = main_instances
-    with pytest.raises(TypeError):
-        nw_v1.get_level(df)  # type: ignore[arg-type]
-    with pytest.raises(TypeError):
-        nw_v1.get_level(ser)  # type: ignore[arg-type]
-    with pytest.raises(TypeError):
-        nw_v1.get_level(lf)  # type: ignore[arg-type]
-
-
-def test_get_level_main_via_v1_from_native(main_instances: MainInstances) -> None:
-    df, lf, ser = main_instances
-    df_v1 = nw_v1.from_native(df)
-    # NOTE: Typing doesn't match the behavior following (#3515)
-    _lf_v1 = nw_v1.from_native(lf)  # type: ignore[call-overload]
-    lf_v1: nw_v1.LazyFrame[Any] = _lf_v1
-    ser_v1 = nw_v1.from_native(ser, series_only=True)
-
-    assert nw_v1.get_level(df_v1) == "full"
-    assert nw_v1.get_level(ser_v1) == "full"
-    if lf_v1.implementation.is_polars():
-        assert nw_v1.get_level(lf_v1) == "lazy"
-    else:
-        # Well this is strange?
-        assert nw_v1.get_level(lf_v1) == "full"
 
 
 def test_any_horizontal() -> None:
